+++ gcovr/gcov_parser.py
@@ -52,13 +52,14 @@ from typing import (
     NoReturn,
 )
 
-from .coverage import BranchCoverage, FileCoverage, FunctionCoverage
+from .coverage import BranchCoverage, FileCoverage, FunctionCoverage, CallCoverage
 from .decision_analysis import DecisionParser
 from .merging import (
     MergeOptions,
     get_or_create_line_coverage,
     insert_branch_coverage,
     insert_function_coverage,
+    insert_call_coverage,
 )
 
 
@@ -295,6 +296,9 @@ class ParserFlags(enum.Flag):
     PARSE_DECISIONS = enum.auto()
     """Whether decision coverage shall be generated."""
 
+    PARSE_CALLS = enum.auto()
+    """Whether call coverage shall be generated."""
+
 
 _LineWithError = Tuple[str, Exception]
 
@@ -504,7 +508,22 @@ def _gather_coverage_from_line(
         return state
 
     # ignore unused line types, such as specialization sections
-    elif isinstance(line, (_CallLine, _UnconditionalLine, _BlockLine)):
+    elif isinstance(line, _CallLine):
+        callno, returned = line
+        line_cov = coverage.lines[state.lineno]  # must already exist
+
+        if context.flags & ParserFlags.PARSE_CALLS:
+            insert_call_coverage(
+                line_cov,
+                CallCoverage(
+                    callno=callno,
+                    covered=(returned > 0),
+                ),
+            )
+
+        return state
+
+    elif isinstance(line, (_UnconditionalLine, _BlockLine)):
         return state
 
     else: