+++ gcovr/writer/html.py
@@ -213,6 +213,7 @@ class RootInfo:
         self.directory = None
         self.branches = dict()
         self.decisions = dict()
+        self.calls = dict()
         self.functions = dict()
         self.lines = dict()
         self.files = []
@@ -230,6 +231,7 @@ class RootInfo:
         self.functions = dict_from_stat(stats.function, self._coverage_to_class)
         self.branches = dict_from_stat(stats.branch, self._branch_coverage_to_class)
         self.decisions = dict_from_stat(stats.decision, self._coverage_to_class)
+        self.calls = dict_from_stat(stats.call, self._coverage_to_class)
 
     def add_file(self, cdata, link_report, cdata_fname):
         stats = SummarizedStats.from_file(cdata)
@@ -263,6 +265,13 @@ class RootInfo:
             "class": self._coverage_to_class(stats.decision.percent),
         }
 
+        calls = {
+            "total": stats.call.total,
+            "exec": stats.call.covered,
+            "coverage": stats.call.percent_or("-"),
+            "class": self._coverage_to_class(stats.call.percent),
+        }
+
         display_filename = force_unix_separator(
             os.path.relpath(realpath(cdata_fname), self.directory)
         )
@@ -279,6 +288,7 @@ class RootInfo:
                 lines=lines,
                 branches=branches,
                 decisions=decisions,
+                calls=calls,
                 functions=functions,
             )
         )
@@ -309,12 +319,14 @@ def print_html_report(covdata: CovData, output_file, options):
     medium_threshold_branch = options.html_medium_threshold_branch
     high_threshold_branch = options.html_high_threshold_branch
     show_decision = options.show_decision
+    show_calls = options.show_calls
 
     data = {}
     root_info = RootInfo(options)
     data["info"] = root_info
 
     data["SHOW_DECISION"] = show_decision
+    data["SHOW_CALL"] = show_calls
     data["COVERAGE_MED"] = medium_threshold
     data["COVERAGE_HIGH"] = high_threshold
     data["LINE_COVERAGE_MED"] = medium_threshold_line
@@ -457,6 +469,7 @@ def print_html_report(covdata: CovData, output_file, options):
             cdata.branch_coverage(), branch_coverage_class
         )
         data["decisions"] = dict_from_stat(cdata.decision_coverage(), coverage_class)
+        data["calls"] = dict_from_stat(cdata.call_coverage(), coverage_class)
 
         data["source_lines"] = []
         currdir = os.getcwd()
@@ -533,6 +546,7 @@ def dict_from_stat(
 def source_row(lineno, source, line_cov):
     linebranch = None
     linedecision = None
+    linecall = None
     linecount = ""
     covclass = ""
     if line_cov:
@@ -546,12 +560,14 @@ def source_row(lineno, source, line_cov):
         elif line_cov.is_uncovered:
             covclass = "uncoveredLine"
             linedecision = source_row_decision(line_cov.decision)
+        linecall = source_row_call(line_cov.calls)
     return {
         "lineno": lineno,
         "source": source,
         "covclass": covclass,
         "linebranch": linebranch,
         "linedecision": linedecision,
+        "linecall": linecall,
         "linecount": linecount,
     }
 
@@ -581,6 +597,33 @@ def source_row_branch(branches):
         "taken": taken,
         "total": total,
         "branches": items,
+    }
+
+
+def source_row_call(calls):
+    if not calls:
+        return None
+
+    invoked = 0
+    total = 0
+    items = []
+
+    for call_id in sorted(calls):
+        call = calls[call_id]
+        if call.is_covered:
+            invoked += 1
+        total += 1
+        items.append(
+            {
+                "invoked": call.is_covered,
+                "name": call_id,
+            }
+        )
+
+    return {
+        "invoked": invoked,
+        "total": total,
+        "calls": items,
     }