+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Breaking changes:
 New features and notable changes:
 - Accept `NAN %` which is used in GCOV 7.5.0 instead of an invalid value. (:issue:`651`)
 - New :option:`--json-base` to define a base bath used in JSON reports. (:issue:`656`)
+- New :option:`--calls` to report call coverage: function calls invoked/total (:issue:`666`)
 
 Bug fixes and small improvements:
 - Fix :option:`--html-tab-size` feature. (:issue:`650`)